requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.6.0
django-environ==0.12.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.13.2
